requirements.txt

feedparser==6.0.10
newspaper3k==0.2.8
requests==2.28.1
twilio==7.15.0
python-dotenv
lxml_html_clean